Validate full path of sending base64 encoded image types
Just make sure we can read the various saved image types on the
server side, encode them as base64, and send them to the client
where we can decode them and present links to save them as files.

This may not be the ultimate functionality we want in vcdat, but
it shows we can probably achieve whatever we do want in terms of
saving screenshots on the server and client.